extract TaxiScheduleInquiry from TaxiScheduler
