+++ contribs/taxi/src/main/java/org/matsim/contrib/etaxi/optimizer/assignment/AssignmentETaxiOptimizer.java
@@ -208,14 +208,15 @@ public class AssignmentETaxiOptimizer extends AssignmentTaxiOptimizer {
 		Stream<DvrpVehicle> vehiclesBelowMinSocLevel = fleet.getVehicles()
 				.values()
 				.stream()
-				.filter(v -> isChargingSchedulable((EvDvrpVehicle)v, eScheduler, maxDepartureTime));
+				.filter(v -> isChargingSchedulable((EvDvrpVehicle)v, eScheduler.getScheduleInquiry(),
+						maxDepartureTime));
 
 		// filter least charged vehicles
 		// assumption: all b.capacities are equal
 		List<DvrpVehicle> leastChargedVehicles = PartialSort.kSmallestElements(pData.getSize(),
 				vehiclesBelowMinSocLevel, v -> ((EvDvrpVehicle)v).getElectricVehicle().getBattery().getSoc());
 
-		return new VehicleData(timer.getTimeOfDay(), eScheduler, leastChargedVehicles.stream());
+		return new VehicleData(timer.getTimeOfDay(), eScheduler.getScheduleInquiry(), leastChargedVehicles.stream());
 	}
 
 	// TODO MIN_RELATIVE_SOC should depend on %idle
