+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/fifo/FifoRequestInserter.java
@@ -22,9 +22,9 @@ import java.util.Collection;
 
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.contrib.dvrp.fleet.Fleet;
-import org.matsim.contrib.taxi.passenger.TaxiRequest;
 import org.matsim.contrib.taxi.optimizer.BestDispatchFinder;
 import org.matsim.contrib.taxi.optimizer.UnplannedRequestInserter;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 import org.matsim.contrib.taxi.scheduler.TaxiScheduler;
 import org.matsim.core.mobsim.framework.MobsimTimer;
 import org.matsim.core.router.util.TravelDisutility;
@@ -42,7 +42,8 @@ public class FifoRequestInserter implements UnplannedRequestInserter {
 			TravelDisutility travelDisutility, TaxiScheduler scheduler) {
 		this.fleet = fleet;
 		this.scheduler = scheduler;
-		dispatchFinder = new BestDispatchFinder(scheduler, network, timer, travelTime, travelDisutility);
+		dispatchFinder = new BestDispatchFinder(scheduler.getScheduleInquiry(), network, timer, travelTime,
+				travelDisutility);
 	}
 
 	@Override
